fix(cli): windows ssh call
